requirements.txt
mcp[cli]
python-dotenv
chromadb
langchain
langchain-community
langchain-openai
langchain-chroma
aiogram
tiktoken
openai
asyncio
nest_asyncio